[iOS] introduce accounts_mediator
AccountsTVC has the wrong presentation style from ManageSyncSettings per
UX guidelines; currently it is pushed to the navigation stack where it
should be presented modally.

As part of this fix, a refactoring of accountsTVC is in the process to
follow the MVVM model.

This CL:
* introduces the mediator
* introduces a consumer
* introduces a AccountsModelIdentityDataSource
* wires up the mediator to the TVC in the coordinator
* moves all observers from the TVC to the mediator
* updates unittests to use the mediator

Bug: 1474803
Change-Id: If13e6b181540ef5b553371d0b6ed620322660237
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5300416
Reviewed-by: Gauthier Ambard <gambard@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265119}