Improve the text of the policy end-date presubmit error
The existing message was:

In policy Foobar: Support on platform chrome.win can only be removed for
version 123. Please remove all references in the code to that policy
since it will not be supported in the current version 124.

The new message is:

In policy Foobar for platform chrome.win: An end-milestone of 134 was
used. But policies are only allowed to be end-dated at versions that
have already branched, currently M123 or before. Please remove all
references in the code to 134, and instead file a bug with a reminder to
add the end milestone after M133 branches.

Bug: 40268638,326507612
Change-Id: I6abe09a2c963c38643a9597e428211c63596c0a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5320520
Commit-Queue: Julian Pastarmov <pastarmovj@chromium.org>
Auto-Submit: Mason Freed <masonf@chromium.org>
Reviewed-by: Julian Pastarmov <pastarmovj@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265129}