🧇 Add SkipFor3p option for Choice screen triggers
This CL adds a feature flag that controls whether the default
search engine choice dialog should be shown when the default search
engine is non-Google.

When this feature flag is enabled, the default search engine choice
dialog is only shown when the default search engine is Google.
When the feature flag is disabled, we return to the pre-existing
checks filtering out certain types of search engines (manual-custom,
distribution-custom, etc).

Bug: 326579197
Change-Id: Ibb9c37623cdbbeb4c7790c1e2e63476a37272122
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5314418
Commit-Queue: Nicolas Dossou-Gbété <dgn@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265178}