[//cc/slim] Hardcode //cc/slim to use GL_TEXTURE_2D as texture target
//cc/slim currently computes the texture target that it passes to
TransferableResource::MakeGpu() by calling
ClientSharedImage::GetTextureTarget(SCANOUT). This logic dates back to
the introduction of this class [1], where it presumably originally came
from here [2]. However, in [2] the texture target is hardcoded to
GL_TEXTURE_2D unless the resource is an overlay candidate, which it
never is in this case (`is_overlay_candidate` is hardcoded to false).
Hence, GL_TEXTURE_2D should always be used here. It *will* in fact
always be used due to the SharedImageFormat in question being one that
will never be in the exception list for needing a platform-specific
texture target (explained in detail here [3]), and hence
ClientSharedImage::GetTextureTarget() returning GL_TEXTURE_2D [4].

This CL hardcodes this use case to use GL_TEXTURE_2D with an explanatory
comment for clarity.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4214432
[2] https://source.chromium.org/chromium/chromium/src/+/main:cc/trees/layer_tree_host_impl.cc;drc=f4a00cc248dd2dc8ec8759fb51620d47b5114090;l=4663
[3] https://chromium-review.googlesource.com/c/chromium/src/+/5317870/comment/050b26e2_4319271f/
[4] https://source.chromium.org/chromium/chromium/src/+/main:gpu/command_buffer/client/client_shared_image.cc;l=142-145?q=ClientSharedImage::GetText&ss=chromium

Bug: 41494843
Change-Id: Iab87b0f565b1cbf090f49b643c4ca6e4616c623c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5323958
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Colin Blundell <blundell@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265180}