[mullet m3] Generate correct main text and labels for every language
When autofill is triggered from the context menu for a non-address
field, the main text and labels are extracted from the string returned
by AutofillProfile::CreateDifferentiatingLabels().

The first address field in this string is the main text, and what is
after represents the label.

Splitting this string is not easy, as different languages have
different separators or no separators at all.

Thus, this CL iterates through all possible address fields and checks
which one is first in the string.

Demo: https://screenshot.googleplex.com/5jGA8QdKFnY4aAF

Bug: 1493361
Change-Id: I31593395efd169740fcd547899fb3d03098b8397
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5224689
Commit-Queue: Theo Cristea <theocristea@google.com>
Reviewed-by: Bruno Braga <brunobraga@google.com>
Reviewed-by: Florian Leimgruber <fleimgruber@google.com>
Cr-Commit-Position: refs/heads/main@{#1265181}