Disable *SnapshotTest.Test on TSan
These tests are very flaky on TSan.
See linked bug: This is likely not an issue with the tests themselves,
but rather with the setup of the global FeatureList. Still, until an
actual fix is found, they have to be disabled to get the bot green.

Bug: 326168468
Change-Id: Ie0b267b05b3cc370fbdcc0ddda9b2f115a2b7b65
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5321603
Auto-Submit: Marc Treib <treib@chromium.org>
Reviewed-by: Victor Vianna <victorvianna@google.com>
Commit-Queue: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265203}