Allow access() of render nodes in video decoding sandbox for Intel/AMD.
A libdrm patch landed that makes drmGetMinorType() call access() with
node paths [1]. drmGetMinorType() is called from drmGetNodeTypeFromFd()
which is called by libva during initialization.

Before this CL, access() was not allowed for the hardware video decoding
sandbox on ChromeOS. Because of that our Lacros and OOPVD tests broke
when the libdrm patch mentioned above landed. This CL allows doing
access() through the broker process on the sysfs paths corresponding to
the render nodes and (on AMD) /dev/dri.

Note that this is still fine with respect to making sure that the
hardware video decoding sandbox is at least as tight as the GPU process
sandbox. For Intel/AMD, the GPU process pre-sandbox also allows
COMMAND_ACCESS [2].

[1] https://chromium.googlesource.com/chromiumos/third_party/libdrm/+/3bc3cca230c5a064b2f554f26fdec27db0f5ead8%5E%21/

[2] https://source.chromium.org/chromium/chromium/src/+/main:content/common/gpu_pre_sandbox_hook_linux.cc;l=655;drc=181f7593e0dabd6b06288ecb3817608d17b42aea

Bug: b:326228696, b:210759684
Test: tast run ${IP} video.Play.vp9_2_hw_oopvd (on hatch and skyrim)
Test: tast run ${IP} video.Play.h264_hw_lacros (on hatch and skyrim)
Change-Id: I124bf6d89c683829cb4bc2fe2df4b67ac39ef62c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5314486
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Reviewed-by: Matthew Denton <mpdenton@chromium.org>
Commit-Queue: Pilar Molina Lopez <pmolinalopez@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265205}