[Autofill] Call SetPreventHighlightingOfAutofilledFields() on all fields
The purpose of SetPreventHighlightingOfAutofilledFields() is during
a preview to "lowlight" the fields that have already been autofilled,
so that the previewed fields stand out.

Before this CL, this only applied to the fields of the form that was
currently being previewed.

The general motivation is lowlighting the fields of the current
form but not those outside of the current form might confuse the
user, especially since the boundaries of forms don't have to
matche what the user perceives as a form.

The concrete motivation is to unblock crrev.com/c/5321681, which
decouples form_util::ApplyFormFields() from a form, which in turn
is necessary for preview to work in some cases with multiple forms in
an iframe.

Bug: 326733459, 1248585, 1247978
Change-Id: I9d7e5568ebfcc5b24e71f0aa2ff9aa3270bbbd5f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318007
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Auto-Submit: Christoph Schwering <schwering@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1265209}