[Autofill] Eliminate initiating element in ApplyFormAction()
This CL removes the parameter of initiating_element parameter
from ApplyFormAction(). The function now assumes that the
currently-focused field is the triggering element instead.

Since the Autofill popup is hidden on focus loss, the two are
the same except in rare circumstances. See the bug for a more
detailed analysis.

The CL is optimized for review. A followup CL simplifies the
function.

Bug: 326733459
Change-Id: Idc7ffca8ab48517b95b458934b6f3908aa37e335
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5321681
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Commit-Queue: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/main@{#1265244}