FaceGaze: Use third_party/ mediapipe files
https://crrev.com/c/5302076 added mediapipe files to third_party/.
This change pulls those files into the FaceGaze build so that
we can run FaceGaze without a custom build argument. Doing this allows
devs/testers to use FaceGaze on a canary build without having to
deploy a custom version of Chrome. This opens FaceGaze up to UXR
possibilities and overall simplifies the development flow.

A few changes worth noting:
1. We use path mappings to point to third_party/.../vision.d.ts to
ensure the TS build works
2. Files needed at runtime are copied from third_party/ to the
accessibility_common extension directory
3. We remove several files from mediapipe_task_vision.tar.xz since
they are now checked-in to third_party/. We also update the .wasm file
that is contained within the archive to the latest version so that it
stays in sync with the .js file in third_party/.

Bug: b:309121742
Change-Id: I602513ba7419626a084707c4c3c0bf1d59dab7e3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5232013
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Reviewed-by: Rebekah Potter <rbpotter@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265245}