[gbb-cct] Move getting surface renderer to background thread.
In local debugging I see this taking in excess of 100ms. It doesn't have to happen on the UI thread, so it makes sense to move it to a background thread. And now we're doing it on background thread it can happen immediately after data is loaded, rather than waiting for trigger conditions (e.g. not in motion) to be met.

Bug: 326595816
Bug: b/319618660

Change-Id: I3b7d3c50a81fd287e7289675d96c2d275b0c1999
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5320904
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Auto-Submit: Edmund Wright <edmundw@google.com>
Commit-Queue: Jinsuk Kim <jinsukkim@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265247}