[UPMLocalPwd] Triggering code for the post migration sheet
Post migration sheet should be triggered on the next startup after the
password migration has finished.
A new pref is added to indicate that the sheet should be shown. It's
false by default and should be set to true by the migration algorithm.
When the sheet is shown, the pref will be set to false and the sheet
will not show up again. That will be done in the follow-up CL on the
Java side because we want to be sure that the sheet UI was actually
shown.

Bug: b/321218513
Change-Id: I397a9af9081f23cc2c749427e0a3eb1694c77c69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5317654
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Ivana Žužić <izuzic@google.com>
Cr-Commit-Position: refs/heads/main@{#1265263}