views: Add some crash keys
It seems that in `View::OrphanLayers()`, `GetLayersInOrder()`
may return a layer that is not a sibling of the current view's
layer, thus trying to remove it from the view's parent layer
results in a crash.

It's hard to know which layer/view/widget it is due to ICF
(Identical Code Folding), so adding these ccrash keys may
allow us to have some insights.

Bug: b/319941708
Change-Id: Ia6642368b5e653e56001d9457d60b0689a7496db
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5322092
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Commit-Queue: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265276}