Mark web test slow on Mac
The test flakes with timeouts. The test results[1] for
mac13-arm64-rel-tests show that even when it passes, it almost always
takes greater than 5 seconds to run.
Although the bug was filed specifically for that builder, the other
results[2] on Mac show a similar pattern of being slow.

Bug: 326619067

[1] https://ci.chromium.org/ui/test/chromium%3Aci/ninja%3A%2F%2F%3Ablink_wpt_tests%2Fexternal%2Fwpt%2Fcss%2Fcss-scroll-snap-2%2Fsnapchanging%2Fsnapchanging-on-user-root-scroll.tentative.html?q=VHASH%3Ad66f84fc0745c58a
[2] https://ci.chromium.org/ui/test/chromium/ninja%3A%2F%2F%3Ablink_wpt_tests%2Fexternal%2Fwpt%2Fcss%2Fcss-scroll-snap-2%2Fsnapchanging%2Fsnapchanging-on-user-root-scroll.tentative.html

Change-Id: Ifab2c08cec06f118125440c60d6da66802dabd02
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5319622
Commit-Queue: David Awogbemila <awogbemila@chromium.org>
Reviewed-by: Mustaq Ahmed <mustaq@chromium.org>
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265300}