[Protected Audiences] Try to fix JoinInterestGroupNonObjectJSONFields.
InterestGroupBrowserTest.JoinInterestGroupNonObjectJSONFields is very
flaky, flaking several times an hour. This CL switches it over to using
RunAuctionAndWaitForUrl(), which gives it a longer worklet timeout.

Unclear if this is why the test is flaking - if this doesn't fix it,
we may have to disable the test.

Bug: 324713537
Change-Id: I03ff76c39f6af0d1f7308f33c1738306255938cb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5308318
Reviewed-by: Caleb Raitto <caraitto@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265310}