Better handle focus changes in task title textfield
Fixes a crash when tabbing out of the task title textfield (when editing
or adding the task). The crash was occurring because the focus manager
is not handling the case where the blurred view gets deleted mid focus
change. To avoid this, the task UI removal is scheduled asynchronously.

Keeps the task view in editing state if the focus
moves to the "Edit in browser" button (so the button can be focused from
the textfield). The logic to exit the edit task state is adapted to keep
the view in edit state as long as either the "Edit in browser" or the
task title textfield are focused. Deactivating the textfield still
commits the task changes.

Don't disable check button while saving an edited, or creating task, to
allow focus move from the title textfield to the check button (as
disabling the check button will make it unfocusable).

BUG=b/324409607,b/324544407,b/321789324

Change-Id: I9cdf22d664e8a2cc21666f71457ef4e05a4b5fea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318622
Reviewed-by: Artsiom Mitrokhin <amitrokhin@chromium.org>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265322}