[OOPIF PDF] Introduce PDFExtensionTestBase::LoadPdfGetExtensionHost()
Add PDFExtensionTestBase::LoadPdfGetExtensionHost(), which is a shortcut
function to load full page PDFs and get the PDF extension host at the
same time.

Note that if this helper method fails, then the failure is slightly
obscured. It can be unclear if the PDF failed to load or if it failed to
get the PDF extension host. This is a tradeoff to reducing repetitive
code.

Bug: 324636880
Change-Id: Ia4dc8d475dddc08707ca12e7be76f9563ea2c5f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318620
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Andy Phan <andyphan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265325}