[views_ax] Using new GetIsIgnored in IsAccessibilityFocusable
This CL modifies the new getter for IsIgnored to also take into
account if the state has been set through the old pipeline.

This is so that while we migrate all the different views in
multiple CLs both systems can coexist and nothing is broken.

The check for the old pipeline will be removed once all views have
been migrated.

Bug: 325137417
Change-Id: I564d2f6b1ba107287d0e8555da0e555c0887840f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5320749
Commit-Queue: Javier Contreras <javiercon@microsoft.com>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1265338}