Refactor VrUiHost lifetime
Currently, the VrUiHost is intended to be created when a device runtime
is added, and will live as long as that runtime does. However, this adds
a lot of additional complexity to the VrUiHost's internal state. It also
leads to a somewhat weird lifetime management, where the
ChromeXrIntegrationClient is responsible for creating and maintaining a
separate reference to the VrUiHost code. Further, the plumbing to allow
a VrUiHost to be created only exists for the windows-specific
IsolatedXrDeviceProvider, meaning that to enable overlays on Android, we
have to find an alternate route for creating a VrUiHost.

This change moves the VrUiHost to be created/destroyed alongside the
session. This allows for simplifying some of the logic in the class
since it doesn't need to worry about existing across the span of
multiple sessions. This also removes the need for two of the
BrowserXRRuntime observer methods that it was subscribing to, as the
things they were used for can now be done during construction (via new
arguments) and destruction. The remaining BrowserXRRuntime observer
method that VrUiHostImpl used can simply be moved to the class
definition, since it was the only class that overrode that method. This
further contributes to simplifying the lifespan for the VrUiHostImpl.

Finally, since the VrUiHostImpl is now created alongside the session,
we no longer need the XRCompositorHost to provide the ImmersiveOverlay
to the session and can simply supply the ImmersiveOverlay directly when
creating the session.

This change also defines the VrUiHost base class and creation method for
all platforms, since future work in this area to enable Overlays on
Android will require that. There are still a couple of Windows-specific
things that the VrUiHostImpl and their children do (apart from simply
needing an Android-compatible GraphicsDelegate), so for the time being,
Chrome's actual implementation of that method is left to be basically
windows-specific. This will minimize the need for future changes, since
the BrowserXrRuntimeImpl will already need to handle the VrUiHost not
existing, since there are currently no plans to create one for every
runtime type (e.g. ARCore).

Bug: 40901055
Change-Id: I2bbb7b085c689eab6b4e8861d41b44aff627baeb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5304598
Reviewed-by: Chris Bookholt <bookholt@chromium.org>
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Auto-Submit: Alexander Cooper <alcooper@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265368}