[Protected Audiences] Deflake some InterestGroupBrowserTests.
navigator.runAdAuction() has default script timeouts of 50 ms. This is
not long enough on memory bots (and timing may be a bit tight on debug
builds, too, or even on release bots).

This CL migrates InterestGroupBrowserTests that use maybePromise()
from calling navigator.runAdAuction() directly to using
RunAuctionAndWait() and RunAuctionAndWaitForUrl().

This CL also introduces MaybePromiseFunction() to make the tests a
little simpler, and does some minor cleanup of the tests it touches.

Bug: 324713537
Change-Id: I5b6898da28fe604dbd9f8ca2b52997bd7b04cdfa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5321663
Reviewed-by: Caleb Raitto <caraitto@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265370}