launcher: Fix crash in TopIconAnimationView
Don't update icon background layer color if layer was not created.
Depending on whether the animation view animates an icon that's shown
within the folder icon or not, the icon background may not have a layer.
When animating a view that doesn't animate into the folder icon, the
background color does not animate differently from the rest of the icon
so no extra layer is needed, so the layer does not exist.

BUG=b/325103208

Change-Id: I735d0dddd8f98f6142caf43d5e89fb5ef1a0f367
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5324338
Reviewed-by: Ana Salazar <anasalazar@chromium.org>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265382}