Fix CpuProbeManager crash on non-desktop platforms
CpuProbeManager is crashing because PressureManagerImpl is incorrectly
checking if cpu_probe_manager_ is null [1], instead of the CpuProbe that
cpu_probe_manager_ owns. As a result, on non-desktop platforms where
CpuProbe is nullptr, the check incorrectly passes and so
CpuProbeManager::EnsureStarted() is called, which crashes attempting to
call CpuProbe::StartSampling on the nullptr.

This CL changes PressureManagerImpl to use a factory to instantialize
CpuProbeManager, that will return a nullptr when the underlying CpuProbe
is nullptr. This way PressureManagerImpl doesn't need to know about
CpuProbe, and can keep using the current check.

[1] https://source.chromium.org/chromium/chromium/src/+/main:services/device/compute_pressure/pressure_manager_impl.cc;l=66

Bug: 326491562
Change-Id: I1bfcd1026c62188147759d87274d55ca32d7a438
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5319960
Reviewed-by: Raphael Kubo Da Costa <raphael.kubo.da.costa@intel.com>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Commit-Queue: Charles Meng <charlesmeng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265408}