Use scope-based InSequence in test expectations
These tests were doing `.InSequence(s)` on every expectations to
check API call ordering. This would be useful if multiple independent
sequences had to be checked in parallel, but it's not the case here.
It's less verbose to use a stack-based `InSequence` object. It's
also safer because it's easy to forget to include an `.InSequence(s)`
on some expectations (which was the case in a few tests, and which led to incorrectly ordered expectations).

Bug: 40249439, 326503102
Change-Id: I929c9bb62e04bd216159633d17555719a590a5f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5321661
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265426}