[Protected Audiences] Deflake some InterestGroupBrowserTests.
navigator.runAdAuction() has default script timeouts of 50 ms. This is
not long enough on memory bots (and timing may be a bit tight on debug
builds, too, or even on release bots).

This CL updates InterestGroupBrowserTests that have hard-coded buyer
and seller timeouts that are less than 5 seconds that are not
supposed to be hit are at least 5 seconds.

Bug: 324713537
Change-Id: I716638594667b73968c00a23981a851dd8aea2c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5323159
Reviewed-by: Caleb Raitto <caraitto@chromium.org>
Commit-Queue: mmenke <mmenke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265428}