Fix setTransform/resetTransform inside canvas layers
Canvas layers are recorded into a side recording until they are closed.
This is done so that we could flush the canvas up until the first
beginLayer since unclosed layers can't be flushed. When layers are
closed, the side recording gets inserted into the main one by using a
DrawRecordOp.

Now, DrawRecordOp was designed to draw recordings as if they had their
own matrix transform root. That is, if the drawn record contains a
SetMatrixOp, that matrix will be applied over the transform the canvas
had when DrawRecordOp was called. This is incompatible with how Canvas
layers work. Calling SetMatrixOp inside a layer should override the
global canvas matrix.

This CL reconciles the two behaviors by adding a `local_ctm` boolean
parameter in DrawRecordOp. When local_ctm is true, the record has a
local matrix stack which won't leak into the drawn canvas and will have
record-local SetMatrixOp. When local_ctm is false, the matrix
operations in the recording behaves as if they were part of the parent
recording.

Bug: 40249439
Fixed: 326503102
Change-Id: I1aa6027124e66b4bb362b473295396d7d6b9ada5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5321885
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265454}