utr: Add basic recipes.py invocation for running the UTR recipe
This hooks up the UTR script to an actual `recipes.py run` invocation.
For now, it invokes recipes.py a single time, and plumbs the recipe
stdout to the UTR's stdout. This means the user will get spammed
with a bunch of recipe junk output. Cleaning and filtering that will
be in a follow-up.

Bug: 41492688
Change-Id: I41d016d881f986ceff8274395a2cba3cb0e24481
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5322459
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Reviewed-by: Struan Shrimpton <sshrimp@google.com>
Cr-Commit-Position: refs/heads/main@{#1265462}