Copy GIFs to clipboard with known size
Adding this to the insertion logic (`ui::TextInputClient::InsertImage`)
will be difficult - similar to adding alt text for inserted gifs - so
this CL only adds the size to GIFs copied to clipboard.

Note that this will stop being completely accurate once we start
inserting full-sized GIFs, as this is currently using the dimensions of
the preview.

Bug: b:322900073
Change-Id: Id2144f1f41feec0a93a0e09a9c76c40685b25916
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5318855
Reviewed-by: Darren Shen <shend@chromium.org>
Commit-Queue: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265467}