Support Feature Tile Dynamic Sizing
Have `FeatureTile` set a `MaximumFlexSizeRule` of `kUnbounded` so it can
dynamically stretch up to the maximum size of the container it is in,
instead of setting a fixed preferred size.

Since the tiles in QS do have a fixed size, set this preferred width and
update the `MaximumFlexRule` to use `kPreferred` from the side of
`FeatureTilesContainerView`. Also adjust pixel tests that don't recreate
the complete `QuickSettingsView` to match its sizing.

To see an example of dynamic sizing, see "display_detailed_view.cc".

Bug: b:295179694
Change-Id: I8cd55bb37a5359c9db3f39a2568e3376a01c9b86
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5319337
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Commit-Queue: Kevin Radtke <kradtke@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1265493}